Refactor: 파일 경로 변경 및 코드 수정 (#6)
