+++ src/components/features/bicycles/category/index.ts
@@ -2,9 +2,5 @@ export { default as BicycleSidebar } from "./bicycle-sidebar";
 export { default as BicycleCard } from "./bicycle-card";
 export { default as BicycleCategoryContent } from "./bicycle-category-content";
 export { default as BicycleSubcategoryTab } from "./bicycle-subcategory-tab";
-export { default as DynamicBreadcrumb } from "./dynamic-breadcrumb";
 export { default as MobileBicycleSidebar } from "./mobile-bicycle-sidebar";
 export { default as BicycleTabsClient } from "./bicycle-tabs-client";
-
-// Zustand 스토어 export
-export { useBicycleFilterStore } from "@/stores/bicycle-filter-store";
